rag-processing/requirements.txt
# PDF processing
pymupdf>=1.24.0

# Text processing
nltk>=3.8.0

# Embeddings
sentence-transformers>=2.2.0
onnxruntime>=1.16.0

# HTTP client
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
